Fix a case where the result payload is lost in MADSchedule
Passing the payload in through step doesn't work, the whole result
naming in _step is misleading. The payload must be passed in through
ctx._result otherwise it can be lost if the ctx is immediately
deferred. This could happen if we over commited outstanding MADs,
which appears to be another woopies due to the semantic change in
steps result argument over time.