The pkey_index returned by umad can be invalid for SMPs
Presumably since there is no validation of the pkey in the SMP path
we've seen that mthca (at least) likes to randomly return garbage in
this field.

Seems like a kernel/driver/firmware bug, but work around it by
ignoring pkey_index for QP 0. This results in all SMP recvs being
tagged with the default pkey which seems to be reasonable from a spec
perspective.