Search for Limited and Full versions of every pkey
If the SA uses a limited default pkey on an EP we need to find that instead of
requiring the full pkey. Generally speaking various places in IBA will return
a full or limited key expecting the node to do this search. We try to match
exactly first.

This also defers looking up the SA default pkey until we actually need
to use it.