Add scripting to build python-rdma through docker
This makes it very easy for anyone to reproduce the various distribution
packages. docker is used to construct an image which has exactly the
config required to execute the build, then the build is run inside that
container and the result copied out.

docker is very fast at setting up containers, so there is almost no
penalty.